New file for function utils
